+++ metadata.rb
@@ -1,7 +1,7 @@
-name             'rhr-entropy'
+name             'rhr_entropy'
 maintainer       'RHR'
 maintainer_email 'JeromeLJordan@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures rhr-entropy'
+description      'Installs/Configures rhr_entropy'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '1.0.0'